Change PIC flag
Something seems off with this, or I'm just not understanding; in another
project, the linker gave me no errors when statically linking this
library (at least not about PIC), but now it does and I don't know what
to think.
It only occurs when I link that one with '-shared'/'-s'/other so flags.