v0.0.2c, 'finished' new log styling feature
It's not where I would like it, and I'm not sure it ever will be,
because in the case of printing with no maximum log, I can't figure out
how to style specific parts of the input format string without
allocating and in someway copying it each time one wishes to log
anything.

It's very much a nuisance, and the only way I can think for it to be
fixed is someway for printf to read only a specified number of bytes
from the format string, rather than reading to the first null-byte.

Requiring the null byte makes it so if the format string given to brrlog
is a string literal (the most common and easy case), I can't do some
slight trickery on the string itself to get printf to play nice (it'll
segfault at runtime for writing read-only memory) so I must allocate and
copy a version that I can manipulate.

The next version 'd' will do things more like they were in version 'b',
since this version is significantly more complicated and if I'm going to
be allocating and copying either way, I might as well leave everything
else as simple as possible.

v.0.0.2 will be the next release, with only a few convenience macros so
that parameterizing calls to brrlog is simpler.