v0.0.2/d I've come up with a solution
Albeit, I wish I didn't have to, but it seems that no matter what, for
absolute correctness, I MUST allocate and duplicate format strings when
logging; given that, I've decided to try and minimize the necessary
(re)allocations by having a global buffer that resizes when
appropriate.

It's resizing is implemented like a hash-table, if the styled form of
a given format string surpasses a factor of the current allocated size
(say, 1.5x), we allocate a new size in that factor so that if another
long string is given, reallocation is not necessary. But, to make sure
large formats don't take up lots of memory for too long, if enough
consecutive formats are small enough, the buffer is shrunk to
accommodate.

It likely still needs some thorough testing, there's probably a few
segfaults hidden in there somewhere, but I'll run into those when I run
into them.

In a perfect world, I could just tell printf not to read more than 'n'
bytes from a format string, but this is not a perfect world.