v0.0.3/c: A simpler implementation of `brrpath`
One of the most useful (I personally think) utilities of a path wrapper
type library is the ability to stat paths correctly and simply;
currently, `brrpath_[f]stat` does this, returning a `brrpath_info_t`
struct that contains the `type` of path, its `size` in bytes, some
`flags` signifying extra information, and eventually (maybe) some
`perms` (permissions).
If a path is a symlink, then one of two things is true:
  If the symlink is valid, then the 'size', 'flags', and 'perms' fields
    are for the target of the symlink; the type of the target is stored in
    the latter 4-bits of the 'type' field (mask `0xF0`).
  Else, the symlink is broken, all fields refer to the symlink itself,
    and the `brrpath_flag_broken` is set in `flags`.

As I've said somewhere before, testing `brrpath` will be an utter
nightmare, so don't be surprised if things are broken in really silly
ways on any operating system.