v0.0.3/d: `brrapi` error system
`brrapi` now provides an error message mechanism, allowing `brrtools` to
give a more descriptive message should an error occur, including which
function (as well as file and line information when debugging the
library itself) the error occured in, and any OS-specific error message
if applicable at the time of error.
To get the error message after a `brr*` function call failed, simply
call `brrapi_error_message` with a destination buffer of a specified
size to copy the full message out. Use `brrapi_error_code` and
`brrapi_error_name` to get the generic code & name for the error,
respectively. Most of the library now uses this mechanism.

Also, improved some of the docstrings/behavior of `brrarray`,
specifically `brrarray_[split|partition]` now partition the source array
into any of the output arrays; if one of the outputs is NULL, it is
simply ignored.
Also, `brrarray_sort` no longer errors; doesn't really make sense for a
sorting function to have an error condition, it either sorts or it
doesn't.